refactor: Move package to oauth from login
