+++ src/main/java/gift/oauth/KakaoOauthConfigure.java
@@ -1,4 +1,4 @@
-package gift.login;
+package gift.oauth;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
