+++ src/main/java/gift/oauth/KakaoTokenResponseDTO.java
@@ -1,4 +1,4 @@
-package gift.login;
+package gift.oauth;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
